# apps/backend/user/user-app-api/contracts/openapi/user-app-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/agritech/orders": {
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderViewDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Bad Request"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [400],
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [401],
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Not Found"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [404],
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Conflict"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [409],
                      "example": 409
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Internal Server Error"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [500],
                      "example": 500
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "nrb.sid": []
          },
          {
            "__Host-nrb.sid": []
          }
        ],
        "tags": ["agritech-orders"]
      },
      "get": {
        "operationId": "OrderController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderListDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Bad Request"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [400],
                      "example": 400
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [401],
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Not Found"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [404],
                      "example": 404
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Conflict"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [409],
                      "example": 409
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "title", "status"],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "example": "about:blank",
                      "description": "A URI reference identifying the problem type; defaults to about:blank when omitted.",
                      "enum": ["about:blank"]
                    },
                    "title": {
                      "type": "string",
                      "description": "A short human-readable summary of the problem type.",
                      "example": "Internal Server Error"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "description": "The HTTP status code generated by the origin server for this occurrence.",
                      "enum": [500],
                      "example": 500
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation specific to this occurrence."
                    },
                    "instance": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "A URI reference identifying this specific occurrence."
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {
            "nrb.sid": []
          },
          {
            "__Host-nrb.sid": []
          }
        ],
        "tags": ["agritech-orders"]
      }
    },
...
  },
  "info": {
    "title": "user-app-api",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nrb.sid",
        "description": "Development/default HTTP session cookie. SESSION_COOKIE_NAME may override the runtime name."
      },
      "__Host-nrb.sid": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-nrb.sid",
        "description": "Production default HTTPS session cookie. SESSION_COOKIE_NAME may override the runtime name."
      }
    },
    "schemas": {
...
      "OrderItemViewDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          },
          "productName": {
            "type": "string"
          },
          "unitPriceUzs": {
            "type": "number"
          },
          "totalUzs": {
            "type": "number"
          }
        },
        "required": ["productId", "quantity", "productName", "unitPriceUzs", "totalUzs"]
      },
      "OrderViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "farmerId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemViewDto"
            }
          },
          "totalAmountUzs": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "processing", "shipped", "delivered", "cancelled"]
          },
          "deliveryAddress": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "userId",
          "farmerId",
          "items",
          "totalAmountUzs",
          "status",
          "deliveryAddress",
          "region",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          }
        },
        "required": ["productId", "quantity"]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "deliveryAddress": {
            "type": "string",
            "example": "Fargona, Quva tumani"
          },
          "region": {
            "type": "string",
            "example": "Farg'ona viloyati"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": ["items", "deliveryAddress", "region"]
      },
      "OrderListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderViewDto"
            }
          }
        },
        "required": ["items"]
      },
...
    }
  }
}